+++ source/extensions/isaacsim.gui.menu/isaacsim/gui/menu/utilities_menu.py
@@ -14,7 +14,10 @@
 # limitations under the License.
 """Utilities menu layout for Isaac Sim."""
 import omni.kit.menu.utils
+from isaacsim.gui.components.menu import make_menu_item_description
 from omni.kit.menu.utils import LayoutSourceSearch, MenuItemDescription, MenuLayout
+
+from .asset_check import AssetCheck
 
 
 class UtilitiesMenuExtension:
@@ -25,8 +28,15 @@ class UtilitiesMenuExtension:
     """
 
     def __init__(self, ext_id: str) -> None:
+        self._asset_check = AssetCheck()
+
         self._menu_placeholder = [MenuItemDescription(name="placeholder", show_fn=lambda: False)]
         omni.kit.menu.utils.add_menu_items(self._menu_placeholder, "Utilities")
+
+        self._menu_items = [
+            make_menu_item_description(ext_id, "Check Default Assets Root Path", self._asset_check.check_assets),
+        ]
+        omni.kit.menu.utils.add_menu_items(self._menu_items, "Utilities")
 
         self.__menu_layout = [
             MenuLayout.Menu(
@@ -38,6 +48,7 @@ class UtilitiesMenuExtension:
                     MenuLayout.Item("Profiler", source="Window/Profiler"),
                     MenuLayout.Item("Statistics"),
                     MenuLayout.Seperator(),
+                    MenuLayout.Item("Check Default Assets Root Path"),
                     MenuLayout.Item("Generate Extension Templates"),
                     MenuLayout.Item("Registered Actions", source="Window/Actions"),
                 ],
@@ -55,4 +66,7 @@ class UtilitiesMenuExtension:
                 menu.shutdown()
         """
         omni.kit.menu.utils.remove_layout(self.__menu_layout)
+        omni.kit.menu.utils.remove_menu_items(self._menu_items, "Utilities")
         omni.kit.menu.utils.remove_menu_items(self._menu_placeholder, "Utilities")
+        self._asset_check.destroy()
+        self._asset_check = None